requirements.txt
requests
beautifulsoup4
selenium
pandas
pyyaml
schedule
python-dotenv
